requirements.txt
Django
djangorestframework
azure-storage-blob
python-dotenv
azure-ai-translation-document==1.0.0
django-environ